iTextSharp-LGPL.4.1.6_Example added
The previous version 4.1.6 has a free license (LGPL) , however the way
of getting the text it's different, so I added an example of how to use
the previous version. Remember to change the references in the solution,
by default, it's using the older version 5.5.9 which can NOT be used for
commercial purposes.